requirements.txt
amqp==2.6.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.2
autopep8==2.0.2
billiard==3.6.4.0
bradocs4py==1.3.1.0
celery==4.4.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
crispy-tailwind==0.5.0
cryptography==38.0.4
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==1.0.0
Django==3.2.10
django-allauth==0.54.0
django-appconf==1.0.5
django-celery-results==2.0.1
django-cpf==0.1.0
django-cpf-cnpj==1.0.0
django-crispy-forms==1.14.0
django-dynamic-preferences==1.12.0
django-filter==23.2
django-fsm==2.8.1
django-fsm-admin==1.2.5
django-fsm-log==3.1.0
django-js-asset==2.1.0
django-modelcluster==6.0
django-mptt==0.14.0
django-polymorphic==3.1.0
django-polymorphic-tree==2.1
django-rest-knox==4.2.0
django-tag-parser==3.2
django-taggit==4.0.0
djangorestframework==3.14.0
Geraldo @ https://github.com/tfs4/geraldo/archive/master.zip
html5lib==1.1
idna==3.4
IMAPClient==2.3.1
kombu==4.6.11
lxml==4.9.1
mgclipboard==0.3.0
mglib==1.2.8
oauthlib==3.2.2
oscrypto==1.3.0
persisting-theory==1.0
Pillow==9.5.0
Py-Cep==1.0.5
pycodestyle==2.10.0
pycparser==2.21
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
PyJWT==2.7.0
pymssql==2.2.7
pyOpenSSL==22.1.0
pypdf==3.12.2
pypng==0.20220715.0
PySIGNFe @ https://github.com/tfs4/PySIGNFe/archive/master.zip
python-bidi==0.4.2
python-decouple==3.1
python3-openid==3.2.0
pytz==2018.3
PyYAML==6.0
qrcode==7.4.2
redis==4.6.0
reportlab==3.6.11
requests==2.31.0
requests-oauthlib==1.3.1
signxml==3.0.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
typing-extensions==4.6.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.3
vine==1.3.0
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.11
django-simple-history==3.2.0
mssql-django
django-storages[azure]
gunicorn